# api-swagger/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.1'
  title: MusicRoomAPI
host: localhost:3001
basePath: /
schemes:
  # tip: remove http to make production-grade
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/api/user/account-validation':
    x-swagger-router-controller: user
    get:
      tags:
        - User
      # description: Route to validate new account
      operationId: accountValidation
      parameters:
        - in: query
          name: token
          type: string
          description: token can be found in your mails
          required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/SuccessResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/user/recover':
    x-swagger-router-controller: user
    post:
      tags:
        - User
      # description: Route to recover account
      operationId: recover
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/RecoverInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/SuccessResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/user/new-password':
    x-swagger-router-controller: user
    get:
      tags:
        - User
      # description: Route to create new password
      operationId: newPassword
      parameters:
        - in: query
          name: token
          type: string
          description: token can be found in your mails
          required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/SuccessResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/user/update-data':
    x-swagger-router-controller: user
    post:
      tags:
        - User
      # description: Route to update simple user data
      operationId: updateData
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UpdateUserDataInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/User'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/user/update-privacy':
    x-swagger-router-controller: user
    post:
      tags:
        - User
      # description: Route to update simple user data
      operationId: updatePrivacy
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UpdatePrivacyInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/User'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/user/link-account':
    x-swagger-router-controller: user
    post:
      tags:
        - User
      # description: Route to update simple user data
      operationId: linkAccount
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/LinkAccountInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/User'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/user/unlink-account':
    x-swagger-router-controller: user
    post:
      tags:
        - User
      # description: Route to update simple user data
      operationId: unlinkAccount
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UnlinkAccountInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/User'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
  '/api/mtv/event':
    x-swagger-router-controller: mtv
    post:
      tags:
        - Music Track Vote
      # description: Route to update simple user data
      operationId: createEvent
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreateEventInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Events'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - Music Track Vote
      # description: Route to update simple user data
      operationId: deleteEvent
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/DeleteEventInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Events'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    patch:
      tags:
        - Music Track Vote
      # description: Route to update simple user data
      operationId: updateData
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UpdateEventDataInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Events'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
  UnlinkAccountInput:
    required:
      - type
    properties:
      type:
        type: string
  LinkAccountInput:
    required:
      - type
      - key
    properties:
      type:
        type: string
      key:
        type: string
  UpdatePrivacyInput:
    required:
      - dataType
      - privacyValue
    properties:
      dataType:
        type: string
      privacyValue:
        type: string
  UpdateUserDataInput:
    required:
      - toChange
      - newValue
    properties:
      toChange:
        type: string
      newValue:
        type: string
  RecoverInput:
    required:
      - email
    properties:
      email:
        type: string
...
  User:
    properties:
      email:
        type: string
      name:
        type: string
      firstname:
        type: string
      avatarUri:
        type: string
      deezer:
        type: boolean
      facebook:
        type: boolean
      google:
        type: boolean
      tags:
        type: array
        items:
          type: string
      privacy:
        type: object
  Events:
    type: array
    items:
      $ref: '#/definitions/Event'
  Event:
    properties:
      id:
        type: string
      name:
        type: string
      canEdit:
        type: boolean
      playlistId:
        type: number
      restriction:
        $ref: '#/definitions/Restriction'
      allowedUsers:
        type: array
        items:
          type: string
      privacy:
        type: string
  Restriction:
    properties:
      isRestricted:
        type: boolean
      location:
        type: string
      maxDistance:
        type: integer
      startDate:
        type: integer
      endDate:
        type: integer
  CreateEventInput:
    required:
      - location
      - name
    properties:
      location:
        type: string
      name:
        type: string
  DeleteEventInput:
    required:
      - location
      - eventId
    properties:
      location:
        type: string
      eventId:
        type: string
  UpdateEventDataInput:
    required:
      - toChange
      - newValue
      - eventId
      - location
    properties:
      toChange:
        type: string
      newValue:
        type: string
      eventId:
        type: string
      location:
        type: string
...
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - error
    properties:
      error:
        type: string